refactored persistance api to org.openhab and removed compatibility layer for persistence services
