+++ bundles/org.openhab.core.model.persistence/src/org/eclipse/smarthome/model/persistence/internal/PersistenceModelManager.java
@@ -16,17 +16,6 @@ import java.util.LinkedList;
 import java.util.List;
 
 import org.eclipse.emf.ecore.EObject;
-import org.eclipse.smarthome.core.persistence.PersistenceManager;
-import org.eclipse.smarthome.core.persistence.PersistenceService;
-import org.eclipse.smarthome.core.persistence.PersistenceServiceConfiguration;
-import org.eclipse.smarthome.core.persistence.SimpleFilter;
-import org.eclipse.smarthome.core.persistence.SimpleItemConfiguration;
-import org.eclipse.smarthome.core.persistence.config.SimpleAllConfig;
-import org.eclipse.smarthome.core.persistence.config.SimpleConfig;
-import org.eclipse.smarthome.core.persistence.config.SimpleGroupConfig;
-import org.eclipse.smarthome.core.persistence.config.SimpleItemConfig;
-import org.eclipse.smarthome.core.persistence.strategy.SimpleCronStrategy;
-import org.eclipse.smarthome.core.persistence.strategy.SimpleStrategy;
 import org.eclipse.smarthome.model.core.EventType;
 import org.eclipse.smarthome.model.core.ModelRepository;
 import org.eclipse.smarthome.model.core.ModelRepositoryChangeListener;
@@ -38,6 +27,17 @@ import org.eclipse.smarthome.model.persistence.persistence.ItemConfig;
 import org.eclipse.smarthome.model.persistence.persistence.PersistenceConfiguration;
 import org.eclipse.smarthome.model.persistence.persistence.PersistenceModel;
 import org.eclipse.smarthome.model.persistence.persistence.Strategy;
+import org.openhab.core.persistence.PersistenceManager;
+import org.openhab.core.persistence.PersistenceService;
+import org.openhab.core.persistence.PersistenceServiceConfiguration;
+import org.openhab.core.persistence.SimpleFilter;
+import org.openhab.core.persistence.SimpleItemConfiguration;
+import org.openhab.core.persistence.config.SimpleAllConfig;
+import org.openhab.core.persistence.config.SimpleConfig;
+import org.openhab.core.persistence.config.SimpleGroupConfig;
+import org.openhab.core.persistence.config.SimpleItemConfig;
+import org.openhab.core.persistence.strategy.SimpleCronStrategy;
+import org.openhab.core.persistence.strategy.SimpleStrategy;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
